Final cleanup.
