+++ Makefile
@@ -18,8 +18,8 @@ lightstep_thrift/constants.go: $(GOPATH)/src/github.com/lightstep/common-go/crou
 	rm -rf lightstep_thrift/reporting_service-remote
 endif
 
-lightstepfakes/fake_recorder.go: options.go
-	counterfeiter -o lightstepfakes/fake_recorder.go options.go Recorder
+lightstepfakes/fake_recorder.go: interfaces.go
+	counterfeiter -o lightstepfakes/fake_recorder.go interfaces.go SpanRecorder
 
 lightstep_thrift/lightstep_thriftfakes/fake_reporting_service.go: lightstep_thrift/reportingservice.go
 	counterfeiter lightstep_thrift/reportingservice.go ReportingService
