+++ grpc_tracer_test.go
@@ -2,125 +2,17 @@ package lightstep_test
 
 import (
 	"encoding/json"
-	"fmt"
-	"reflect"
 	"strings"
 	"time"
-
-	"golang.org/x/net/context"
-	"google.golang.org/grpc"
 
 	. "github.com/lightstep/lightstep-tracer-go"
 	cpb "github.com/lightstep/lightstep-tracer-go/collectorpb"
 	cpbfakes "github.com/lightstep/lightstep-tracer-go/collectorpb/collectorpbfakes"
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
-	"github.com/onsi/gomega/types"
 	ot "github.com/opentracing/opentracing-go"
 	"github.com/opentracing/opentracing-go/log"
 )
-
-type haveKeyValuesMatcher []*cpb.KeyValue
-
-func HaveKeyValues(keyValues ...*cpb.KeyValue) types.GomegaMatcher {
-	return haveKeyValuesMatcher(keyValues)
-}
-
-func (matcher haveKeyValuesMatcher) Match(actual interface{}) (bool, error) {
-	var actualKeyValues []*cpb.KeyValue
-
-	switch v := actual.(type) {
-	case []*cpb.KeyValue:
-		actualKeyValues = v
-	case *cpb.Log:
-		actualKeyValues = v.GetKeyvalues()
-	default:
-		return false, fmt.Errorf("HaveKeyValues matcher expects either a []*KeyValue or a *Log")
-	}
-
-	expectedKeyValues := []*cpb.KeyValue(matcher)
-	if len(expectedKeyValues) != len(actualKeyValues) {
-		return false, nil
-	}
-
-	for i, _ := range actualKeyValues {
-		if !reflect.DeepEqual(actualKeyValues[i], expectedKeyValues[i]) {
-			return false, nil
-		}
-	}
-
-	return true, nil
-}
-
-func (matcher haveKeyValuesMatcher) FailureMessage(actual interface{}) string {
-	return fmt.Sprintf("Expected '%v' to have key values '%v'", actual, matcher)
-}
-
-func (matcher haveKeyValuesMatcher) NegatedFailureMessage(actual interface{}) string {
-	return fmt.Sprintf("Expected '%v' to not have key values '%v'", actual, matcher)
-}
-
-func KeyValue(key string, value interface{}, storeAsJson ...bool) *cpb.KeyValue {
-	tag := &cpb.KeyValue{Key: key}
-	switch typedValue := value.(type) {
-	case int:
-		tag.Value = &cpb.KeyValue_IntValue{int64(typedValue)}
-	case string:
-		if len(storeAsJson) > 0 && storeAsJson[0] {
-			tag.Value = &cpb.KeyValue_JsonValue{typedValue}
-		} else {
-			tag.Value = &cpb.KeyValue_StringValue{typedValue}
-		}
-	case bool:
-		tag.Value = &cpb.KeyValue_BoolValue{typedValue}
-	case float32:
-		tag.Value = &cpb.KeyValue_DoubleValue{float64(typedValue)}
-	case float64:
-		tag.Value = &cpb.KeyValue_DoubleValue{typedValue}
-	}
-	return tag
-}
-
-func startNSpans(n int, tracer ot.Tracer) {
-	for i := 0; i < n; i++ {
-		tracer.StartSpan(string(i)).Finish()
-	}
-}
-
-func attachGrpcSpanListener(fakeClient *cpbfakes.FakeCollectorServiceClient) func() []*cpb.Span {
-	reportChan := make(chan *cpb.ReportRequest)
-	fakeClient.ReportStub = func(context context.Context, reportResponse *cpb.ReportRequest, options ...grpc.CallOption) (*cpb.ReportResponse, error) {
-		select {
-		case reportChan <- reportResponse:
-		case <-time.After(1 * time.Second):
-		}
-		return &cpb.ReportResponse{}, nil
-	}
-
-	return func() []*cpb.Span {
-		timeout := time.After(5 * time.Second)
-		for {
-			select {
-			case report := <-reportChan:
-				if len(report.GetSpans()) > 0 {
-					return report.GetSpans()
-				}
-			case <-timeout:
-				Fail("timed out trying to get spans")
-			}
-		}
-	}
-}
-
-type dummyConnection struct{}
-
-func (*dummyConnection) Close() error { return nil }
-
-func fakeGrpcConnection(fakeClient *cpbfakes.FakeCollectorServiceClient) ConnectorFactory {
-	return func() (interface{}, Connection, error) {
-		return fakeClient, new(dummyConnection), nil
-	}
-}
 
 var _ = Describe("Tracer", func() {
 	var tracer ot.Tracer
@@ -136,9 +28,7 @@ var _ = Describe("Tracer", func() {
 		})
 
 		AfterEach(func() {
-			errChan := make(chan error)
-			go func() { errChan <- CloseTracer(tracer) }()
-			Eventually(errChan).Should(Receive(BeNil()))
+			closeTestTracer(tracer)
 		})
 
 		Context("With default options", func() {
@@ -215,19 +105,13 @@ var _ = Describe("Tracer", func() {
 
 			Describe("CloseTracer", func() {
 				It("Should not explode when called twice", func() {
-					errChan := make(chan error)
-					go func() { errChan <- CloseTracer(tracer) }()
-					Eventually(errChan).Should(Receive(BeNil()))
-
-					go func() { errChan <- CloseTracer(tracer) }()
-					Eventually(errChan).Should(Receive(BeNil()))
+					closeTestTracer(tracer)
+					closeTestTracer(tracer)
 				})
 
 				It("Should behave nicely", func() {
 					By("Not hanging")
-					errChan := make(chan error)
-					go func() { errChan <- CloseTracer(tracer) }()
-					Eventually(errChan).Should(Receive(BeNil()))
+					closeTestTracer(tracer)
 
 					By("Stop communication with server")
 					lastCallCount := fakeClient.ReportCallCount()
