+++ tag_version.sh
@@ -4,5 +4,5 @@ VERSION=$(cat ./VERSION)
 
 cat > version.go <<EOF
 package lightstep
-var TracerVersionValue = "$VERSION"
+const TracerVersionValue = "$VERSION"
 EOF
