+++ version.go
@@ -1,3 +1,2 @@
 package lightstep
-
-var TracerVersionValue = "0.11.0"
+const TracerVersionValue = "0.13.0"
